Fixed before calculations

+++ sdizo2/sdizo2/source/controllers/testControllers/genericTestController/genericTestController.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 
-void genericTestController::saveToFile(int testSize, int time)
+void genericTestController::saveToFile(int testSize, double time)
 {
 	file.open(filename, std::ios::out | std::ios::app);
 	if (file.good())
